SecondaryProgress for RoundCornerProgressBar Finish
+ Add attribute secondaryProgress and secondaryProgressColor
+ Fixed attribute progress and max format from integer to float
+ Change round_corner_layout from LinearLayout to RelativeLayout
+ Add/Update method setProgressColor(), setProgress(), setSecondaryProgress(), getSecondaryProgress() and getSecondaryProgressColor()
+ Update example project

+++ library/src/main/java/com/akexorcist/roundcornerprogressbar/IconRoundCornerProgressBar.java
@@ -170,10 +170,10 @@ public class IconRoundCornerProgressBar extends LinearLayout {
 		}
 
 		if(!isMaxProgressSetBeforeDraw) {
-			max = typedArray.getInt(R.styleable.RoundCornerProgress_max, 0);
+			max = typedArray.getFloat(R.styleable.RoundCornerProgress_max, 0);
 		}
 		if(!isProgressSetBeforeDraw) {
-			progress = typedArray.getInt(R.styleable.RoundCornerProgress_progress, 0);
+			progress = typedArray.getFloat(R.styleable.RoundCornerProgress_progress, 0);
 		}
 
 		typedArray.recycle();
